requirements.txt
chromadb==0.5.0
fitz==0.0.1.dev2
fonttools==4.25.0
ipython==8.12.3
langchain==0.1.16
langchain_community==0.0.34
langchain_openai==0.1.4
langchain_pinecone==0.1.0
langchain_text_splitters==0.0.1
multiprocess==0.70.16
numpy==1.26.4
openai==1.23.6
Pillow==10.3.0
pinecone-client
PyPika==0.48.9
pyzotero==1.5.18
Requests==2.31.0
streamlit==1.33.0
streamlit_pills==0.3.0
torch==2.2.2
typing_extensions==4.11.0
pysqlite3-binary
PyMuPDF
st-chat-message
dotenv